+++ src/app/app.module.ts
@@ -5,17 +5,17 @@ import { MyApp } from './app.component';
 
 import { SetPage } from '../pages/set/set';
 import { TemplatePage } from '../pages/template/template';
-import { AddTemplatePage } from '../pages/add-template/add-template';
-import { TemplateDetailPage } from '../pages/template-detail/template-detail';
-import { AddItemPage } from '../pages/add-item/add-item';
-import { ItemDetailPage } from '../pages/item-detail/item-detail';
+import { AddTemplatePage } from '../pages/template/add-template/add-template';
+import { TemplateDetailPage } from '../pages/template/template-detail/template-detail';
+import { AddItemPage } from '../pages/template/item/add-item/add-item';
+import { ItemDetailPage } from '../pages/template/item/item-detail/item-detail';
 import { TabsPage } from '../pages/tabs/tabs';
 
-import { AddSetPage } from '../pages/add-set/add-set';
-import { SetDetailPage } from '../pages/set-detail/set-detail';
-import { EditSetPage } from '../pages/edit-set/edit-set';
-import { AddRowPage } from '../pages/add-row/add-row';
-import { RowDetailPage } from '../pages/row-detail/row-detail';
+import { AddSetPage } from '../pages/set/add-set/add-set';
+import { SetDetailPage } from '../pages/set/set-detail/set-detail';
+import { EditSetPage } from '../pages/set/edit-set/edit-set';
+import { AddRowPage } from '../pages/set/row/add-row/add-row';
+import { RowDetailPage } from '../pages/set/row/row-detail/row-detail';
 import { StatusBar } from '@ionic-native/status-bar';
 import { SplashScreen } from '@ionic-native/splash-screen';
