+++ src/pages/set/row/row-detail/row-detail.scss
