+++ src/pages/set/row/row-detail/row-detail.ts
