+++ src/pages/set/set-detail/set-detail.html
