+++ src/pages/set/set.ts
@@ -3,9 +3,9 @@ import { NavController, NavParams, ModalController } from 'ionic-angular';
 
 import { SetData } from '../../providers/set-data/set-data';
 
-import { AddSetPage } from '../add-set/add-set';
-import { EditSetPage } from '../edit-set/edit-set';
-import { SetDetailPage } from '../set-detail/set-detail';
+import { AddSetPage } from './add-set/add-set';
+import { EditSetPage } from './edit-set/edit-set';
+import { SetDetailPage } from './set-detail/set-detail';
 import { UtilsService } from '../../services/utils/utils';
 
 /**
