+++ src/pages/template/item/item-detail/item-detail.scss
