+++ src/pages/template/item/item-detail/item-detail.ts
@@ -1,34 +1,40 @@
 import { Component } from '@angular/core';
 import { NavController, NavParams, ViewController } from 'ionic-angular';
-import { UtilsService } from '../../services/utils/utils';
+import { UtilsService } from '../../../../services/utils/utils';
 
 /**
- * Generated class for the AddItemPage page.
+ * Generated class for the ItemDetailPage page.
  *
  * See https://ionicframework.com/docs/components/#navigation for more info on
  * Ionic pages and navigation.
  */
 
 @Component({
-  selector: 'page-add-item',
-  templateUrl: 'add-item.html',
+  selector: 'page-item-detail',
+  templateUrl: 'item-detail.html',
 })
-export class AddItemPage {
+export class ItemDetailPage {
 
-  title: string;
-  type: string;
-  selectItems = [];
   item;
+  title;
+  type;
+  selectItems = [];
   unit;
 
   constructor(public navCtrl: NavController, public navParams: NavParams, public view: ViewController, public utilsService: UtilsService) {
-    let item = {
-      value: ""
-    }
-    this.selectItems.push(item);
   }
 
   ionViewDidLoad() {
+    if (this.navParams.get('item')) {
+      this.item = this.navParams.get('item');
+      this.title = this.navParams.get('item').title;
+      this.type = this.navParams.get('item').type;
+      //if (this.type == "select") {
+        this.selectItems = this.navParams.get('item').items;
+        this.addNullSelectItem();
+      //}
+      this.unit = this.navParams.get('item').unit;
+    }
   }
 
   reorderSelectItems(indexes) {
@@ -38,7 +44,7 @@ export class AddItemPage {
     this.selectItems.splice(indexes.to, 0, element);
   }
 
-  addNullSelectItem(value) {
+  addNullSelectItem() {
     for (var i = 0; i < this.selectItems.length; i++) {
       if (this.selectItems[i].value == "") {
         return;
@@ -61,19 +67,20 @@ export class AddItemPage {
       this.utilsService.showToast('top', "Item can not be saved without a type!");
       return;
     }
+    this.item.title = this.title;
+    this.item.type = this.type;
+
+    // 保存前，删除最后一个空列表项
     this.selectItems.pop();
-    let newItem = {
-      title: this.title,
-      type: this.type,
-      items: this.selectItems,
-      unit: this.unit
-    };
+    this.item.items = this.selectItems;
+    this.item.unit = this.unit;
 
-    this.view.dismiss(newItem);
-
+    this.view.dismiss(this.item);
   }
 
   close() {
+    // 关闭前，删除最后一个空列表项
+    this.selectItems.pop();
     this.view.dismiss();
   }
