+++ src/pages/template/template-detail/template-detail.scss
